apps/middleware-api/requirements.txt
# Web-Framework
fastapi==0.115.0
uvicorn[standard]==0.30.0

# Datenbank
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0

# Validierung & Einstellungen
pydantic==2.9.0
pydantic-settings==2.5.0

# Datei-Upload
python-multipart==0.0.12

# ID-Generierung
python-ulid==2.7.0

# Tests
pytest==8.3.0
pytest-asyncio==0.24.0
httpx==0.27.0
